requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
PyJWT>=2.9.0
bcrypt>=4.2.0
websockets>=13.0
watchdog>=5.0.0
pyyaml>=6.0
claude-agent-sdk>=0.1.49
ptyprocess>=0.7.0
httpx>=0.28.0
tomli>=2.0.1; python_version < "3.11"
